+++ src/app/pin-entry/pin-entry.component.ts
@@ -5,15 +5,16 @@ import { FormsModule } from '@angular/forms';
 import { UiService } from '../ui/ui.service';
 
 @Component({
-    selector: 'app-pin-entry',
-    templateUrl: './pin-entry.component.html',
-    styleUrls: ['./pin-entry.component.scss'],
-    imports: [IonInput, IonModal, IonText, IonButton, IonIcon, FormsModule]
+  selector: 'app-pin-entry',
+  templateUrl: './pin-entry.component.html',
+  styleUrls: ['./pin-entry.component.scss'],
+  imports: [IonInput, IonModal, IonText, IonButton, IonIcon, FormsModule]
 })
 export class PinEntryComponent {
   show = model(false);
   dismissed = output<boolean>();
   correctPin = input<string>('');
+  encrypted = input<boolean>(true);
   message = input<string>('A PIN is required to access this event. This is delivered by email or at greeters.');
   enteredPin = model('');
   private toastController = inject(ToastController);
@@ -30,7 +31,9 @@ export class PinEntryComponent {
   }
 
   async checkPin(): Promise<boolean> {
-    const correctPin = await decryptString(this.correctPin(), 'd1e0fa-b0b0-4b79-a6ca-8ffdf8be88');
+    const correctPin = this.encrypted() ?
+      await decryptString(this.correctPin(), 'd1e0fa-b0b0-4b79-a6ca-8ffdf8be88') :
+      this.correctPin();
     return correctPin == this.enteredPin();
   }